requirements.txt
sqlalchemy==2.0.22
python-dotenv==1.0.0
marshmallow==3.20.1
pytest==7.4.2
black==23.9.1
isort==5.12.0
alembic==1.13.0
